refactor(usage): move env-name resolution next to the CH path
Make the CH path own its uniform contract: `group.value` is always
the display string. Other dim columns are slug-typed in CH and need
no work; `environment_id` is Int64 and gets resolved here. Pulls
the post-process out of the controller so the controller doesn't
need to special-case env_id.
